Restructured code to respective files
