+++ healthScore/doctor_data.py
@@ -1,0 +1,126 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+from django.http import JsonResponse
+from django.utils import timezone
+from datetime import datetime, timedelta
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required
+from django.forms.models import model_to_dict
+import json
+
+from reportlab.lib.pagesizes import letter
+from reportlab.platypus import (
+    SimpleDocTemplate,
+    Paragraph,
+    Spacer,
+    Table,
+    TableStyle,
+    Image,
+)
+from reportlab.lib.styles import getSampleStyleSheet
+from reportlab.lib import colors
+from reportlab.lib.enums import TA_RIGHT
+from reportlab.lib.styles import ParagraphStyle
+
+# To overcame issues with regards to permissions (POST calls will give CSRF errors if the below tag is not used)
+from django.views.decorators.csrf import csrf_exempt
+
+from healthConfig.settings import EMAIL_HOST_USER
+
+from .models import (
+    Appointment,
+    HealthRecord,
+    Hospital,
+    User,
+    HospitalStaff,
+    Post,
+    Comment,
+    HealthHistoryAccessRequest,
+)
+
+from .user_utils import get_health_history_details
+from .forms import PostForm, CommentForm
+from .file_upload import file_upload
+from django.views.decorators.http import require_http_methods
+from django.core.mail import send_mail, EmailMessage
+from .hospital_admin_utils import get_admin_health_history_details
+
+DATE_FORMAT = "%Y-%m-%d"
+APPOINTMENT_TYPE = {
+    "blood_test": "Blood Test",
+    "eye": "Eye Exams",
+    "general": "General Physical",
+    "dermatologist": "Dermatologist",
+    "diabetes_screening": "Diabetes Screening",
+    "dentist": "Dentist",
+    "gynecologist": "Gynecologist",
+    "vaccinations": "Vaccinations",
+}
+
+APPOINTMENT_PROPS = {
+    "blood_test": {
+        "blood_group": "Blood Group",
+        "hemoglobin_count": "Hemoglobin Count",
+        "date": "Date",
+        "platelet_count": "Platelet Count",
+    },
+    "eye": {
+        "cylindrical_power_right": "Cylindrical Power Right",
+        "cylindrical_power_left": "Cylindrical Power Left",
+        "spherical_power_left": "Spherical Power Left",
+        "spherical_power_right": "Spherical Power Right",
+        "date": "Date",
+    },
+    "general": {
+        "blood_pressure": "Blood Pressure",
+        "pulse_rate": "Pulse Rate",
+        "date": "Date",
+    },
+    "dermatologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "diabetes_screening": {
+        "fasting_sugar_level": "Fasting Sugar Level",
+        "random_sugar_level": "Random Sugar Level",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "dentist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "gynecologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "vaccinations": {
+        "name": "Name",
+        "type": "Vaccination Type",
+        "dose_2": "Dose 2",
+        "date": "Dose 2 Date",
+    },
+}
+
+
+
+@login_required(login_url="/")
+def get_patients(request):
+    patients = list(User.objects.filter(is_patient=True).values())
+    return JsonResponse({"patients": patients})
+
+
+@login_required(login_url="/")
+def get_doctor_details(request, doctor_id):
+    doctor = HospitalStaff.objects.filter(id=doctor_id).first()
+    user_detail = list(User.objects.filter(id=doctor.userID).values())
+    return JsonResponse({"user": user_detail})
+
+
+@login_required(login_url="/")
+def get_patient_details(request, patient_id):
+    patient_detail = list(User.objects.filter(id=patient_id).values())
+    return JsonResponse({"user": patient_detail})
