+++ healthScore/healthcare_data.py
@@ -1,0 +1,288 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+from django.http import JsonResponse
+from django.utils import timezone
+from datetime import datetime, timedelta
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required
+from django.forms.models import model_to_dict
+import json
+
+from reportlab.lib.pagesizes import letter
+from reportlab.platypus import (
+    SimpleDocTemplate,
+    Paragraph,
+    Spacer,
+    Table,
+    TableStyle,
+    Image,
+)
+from reportlab.lib.styles import getSampleStyleSheet
+from reportlab.lib import colors
+from reportlab.lib.enums import TA_RIGHT
+from reportlab.lib.styles import ParagraphStyle
+
+# To overcame issues with regards to permissions (POST calls will give CSRF errors if the below tag is not used)
+from django.views.decorators.csrf import csrf_exempt
+
+from healthConfig.settings import EMAIL_HOST_USER
+
+from .models import (
+    Appointment,
+    HealthRecord,
+    Hospital,
+    User,
+    HospitalStaff,
+    Post,
+    Comment,
+    HealthHistoryAccessRequest,
+)
+
+from .user_utils import get_health_history_details
+from .forms import PostForm, CommentForm
+from .file_upload import file_upload
+from django.views.decorators.http import require_http_methods
+from django.core.mail import send_mail, EmailMessage
+from .hospital_admin_utils import get_admin_health_history_details
+
+DATE_FORMAT = "%Y-%m-%d"
+APPOINTMENT_TYPE = {
+    "blood_test": "Blood Test",
+    "eye": "Eye Exams",
+    "general": "General Physical",
+    "dermatologist": "Dermatologist",
+    "diabetes_screening": "Diabetes Screening",
+    "dentist": "Dentist",
+    "gynecologist": "Gynecologist",
+    "vaccinations": "Vaccinations",
+}
+
+APPOINTMENT_PROPS = {
+    "blood_test": {
+        "blood_group": "Blood Group",
+        "hemoglobin_count": "Hemoglobin Count",
+        "date": "Date",
+        "platelet_count": "Platelet Count",
+    },
+    "eye": {
+        "cylindrical_power_right": "Cylindrical Power Right",
+        "cylindrical_power_left": "Cylindrical Power Left",
+        "spherical_power_left": "Spherical Power Left",
+        "spherical_power_right": "Spherical Power Right",
+        "date": "Date",
+    },
+    "general": {
+        "blood_pressure": "Blood Pressure",
+        "pulse_rate": "Pulse Rate",
+        "date": "Date",
+    },
+    "dermatologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "diabetes_screening": {
+        "fasting_sugar_level": "Fasting Sugar Level",
+        "random_sugar_level": "Random Sugar Level",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "dentist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "gynecologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "vaccinations": {
+        "name": "Name",
+        "type": "Vaccination Type",
+        "dose_2": "Dose 2",
+        "date": "Dose 2 Date",
+    },
+}
+
+
+
+
+def get_facility_doctors(request):
+    if request.user.is_authenticated:
+        user_hospital_staff_entry = get_object_or_404(
+            HospitalStaff, userID=request.user.id
+        )
+        hospital_id = user_hospital_staff_entry.hospitalID.id
+
+        staff_members = HospitalStaff.objects.filter(
+            hospitalID=hospital_id, admin=False
+        )
+
+        staff_data = []
+        for staff in staff_members:
+            try:
+                user = User.objects.get(id=staff.userID)
+                staff_data.append(
+                    {
+                        "id": user.id,
+                        "name": user.name,
+                        "email": user.email,
+                        "contactInfo": staff.contactInfo,
+                        "specialty": staff.specialization,
+                        "is_active": user.is_active,
+                    }
+                )
+            except User.DoesNotExist:
+                continue
+
+        return JsonResponse({"data": staff_data}, safe=False)
+
+    return JsonResponse({"error": "Unauthorized"}, status=401)
+
+
+def get_facility_admins(request):
+    if request.user.is_authenticated:
+        user_hospital_staff_entry = get_object_or_404(
+            HospitalStaff, userID=request.user.id
+        )
+        hospital_id = user_hospital_staff_entry.hospitalID.id
+
+        staff_members = HospitalStaff.objects.filter(hospitalID=hospital_id, admin=True)
+
+        staff_data = []
+        for staff in staff_members:
+            try:
+                user = User.objects.get(id=staff.userID)
+                staff_data.append(
+                    {
+                        "id": user.id,
+                        "name": user.name,
+                        "email": user.email,
+                        "contactInfo": staff.contactInfo,
+                        "specialty": staff.specialization,
+                        "is_active": user.is_active,
+                    }
+                )
+            except User.DoesNotExist:
+
+                continue
+
+        return JsonResponse({"data": staff_data}, safe=False)
+
+    return JsonResponse({"error": "Unauthorized"}, status=401)
+
+
+def hospital_staff_directory(request):
+    context = {
+        "get_facility_doctors_url": "api/get-facility-doctors/",
+        "get_facility_admins_url": "api/get-facility-admins/",
+    }
+    return render(request, "healthcare_facility.html", context)
+
+
+@login_required(login_url="/")
+@csrf_exempt
+def add_healthcare_staff(request):
+    if request.user.is_authenticated and request.method == "POST":
+        email = request.POST.get("email")
+        fullname = request.POST.get("fullname")
+        contactInfo = request.POST.get("contactInfo")
+        is_admin = int(request.POST.get("is_admin"))
+        specialization = request.POST.get("specialization")
+
+        user_hospital_staff_entry = get_object_or_404(
+            HospitalStaff, userID=request.user.id
+        )
+        hospital_id = user_hospital_staff_entry.hospitalID.id
+
+        context = {"error_message:": ""}
+
+        if User.objects.filter(email=email).exists():
+            user = User.objects.get(email=email)
+            if user.is_patient:
+                context["error_message"] = (
+                    "A patient account already exists with this email"
+                )
+            elif user.is_staff:
+                context["error_message"] = (
+                    "An admin account already exists with this email"
+                )
+            else:
+                context["error_message"] = (
+                    "A healthcare worker account already exists with this email"
+                )
+
+            return render(request, "healthcare_facility.html", context)
+
+        user_fields = {
+            "email": email,
+            "password": "dummy_password",  # healthcare worker will have to reset the password on first login
+            "name": fullname,
+            "contactInfo": contactInfo,
+        }
+
+        hospital = get_object_or_404(Hospital, id=hospital_id)
+
+        user = None
+        if is_admin:
+            user = User.objects.create_staff(**user_fields)
+        else:
+            user = User.objects.create_healthcare_worker(**user_fields)
+
+        HospitalStaff.objects.create(
+            hospitalID=hospital,
+            admin=is_admin,
+            name=fullname,
+            specialization=specialization,
+            contactInfo=contactInfo,
+            userID=user.id,
+        )
+
+        return redirect("hospital_staff_directory")
+
+    return JsonResponse({"error": "Unauthorized"}, status=401)
+
+
+@login_required(login_url="/")
+@csrf_exempt
+def deactivate_healthcare_staff(request):
+    if request.user.is_authenticated and request.method == "PUT":
+        updated_data = json.loads(request.body)
+        user_ids = updated_data.get("user_ids", [])
+
+        for user_id in user_ids:
+            user = get_object_or_404(User, id=user_id)
+            if not user.is_patient:
+                user.is_active = False
+                user.save()
+
+        return JsonResponse(
+            {"message": "Healthcare staff deactivated successfully"}, status=200
+        )
+
+    return JsonResponse({"error": "Unauthorized"}, status=401)
+
+
+@login_required(login_url="/")
+@csrf_exempt
+def activate_healthcare_staff(request):
+    if request.user.is_authenticated and request.method == "PUT":
+        updatedData = json.loads(request.body)
+        user_id = updatedData.get("user_id")
+
+        user = get_object_or_404(User, id=user_id)
+
+        if user.is_patient:
+            return JsonResponse(
+                {"error": "Patient's account cannot be edited"}, status=400
+            )
+
+        user.is_active = True
+        user.save()
+        return JsonResponse(
+            {"message": "Healthcare staff activated successfully"}, status=200
+        )
+
+    return JsonResponse({"error": "Unauthorized"}, status=401)
+
