+++ healthScore/patient_view_records.py
@@ -1,0 +1,278 @@
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
+from django.http import JsonResponse
+from django.utils import timezone
+from datetime import datetime, timedelta
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required
+from django.forms.models import model_to_dict
+import json
+
+from reportlab.lib.pagesizes import letter
+from reportlab.platypus import (
+    SimpleDocTemplate,
+    Paragraph,
+    Spacer,
+    Table,
+    TableStyle,
+    Image,
+)
+from reportlab.lib.styles import getSampleStyleSheet
+from reportlab.lib import colors
+from reportlab.lib.enums import TA_RIGHT
+from reportlab.lib.styles import ParagraphStyle
+
+# To overcame issues with regards to permissions (POST calls will give CSRF errors if the below tag is not used)
+from django.views.decorators.csrf import csrf_exempt
+
+from healthConfig.settings import EMAIL_HOST_USER
+
+from .models import (
+    Appointment,
+    HealthRecord,
+    Hospital,
+    User,
+    HospitalStaff,
+    Post,
+    Comment,
+    HealthHistoryAccessRequest,
+)
+
+from .user_utils import get_health_history_details
+from .forms import PostForm, CommentForm
+from .file_upload import file_upload
+from django.views.decorators.http import require_http_methods
+from django.core.mail import send_mail, EmailMessage
+from .hospital_admin_utils import get_admin_health_history_details
+
+DATE_FORMAT = "%Y-%m-%d"
+APPOINTMENT_TYPE = {
+    "blood_test": "Blood Test",
+    "eye": "Eye Exams",
+    "general": "General Physical",
+    "dermatologist": "Dermatologist",
+    "diabetes_screening": "Diabetes Screening",
+    "dentist": "Dentist",
+    "gynecologist": "Gynecologist",
+    "vaccinations": "Vaccinations",
+}
+
+APPOINTMENT_PROPS = {
+    "blood_test": {
+        "blood_group": "Blood Group",
+        "hemoglobin_count": "Hemoglobin Count",
+        "date": "Date",
+        "platelet_count": "Platelet Count",
+    },
+    "eye": {
+        "cylindrical_power_right": "Cylindrical Power Right",
+        "cylindrical_power_left": "Cylindrical Power Left",
+        "spherical_power_left": "Spherical Power Left",
+        "spherical_power_right": "Spherical Power Right",
+        "date": "Date",
+    },
+    "general": {
+        "blood_pressure": "Blood Pressure",
+        "pulse_rate": "Pulse Rate",
+        "date": "Date",
+    },
+    "dermatologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "diabetes_screening": {
+        "fasting_sugar_level": "Fasting Sugar Level",
+        "random_sugar_level": "Random Sugar Level",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "dentist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "gynecologist": {
+        "care_received": "Care Received",
+        "second_visit": "Second Visit Required",
+        "date": "Date",
+    },
+    "vaccinations": {
+        "name": "Name",
+        "type": "Vaccination Type",
+        "dose_2": "Dose 2",
+        "date": "Dose 2 Date",
+    },
+}
+
+
+@login_required(login_url="/")
+def view_health_history_requests(request):
+    if request.user.is_patient:
+        zipped_details = get_health_history_details(request=request)
+        return render(request, "view_requests.html", {"zipped_details": zipped_details})
+    redirect("homepage")
+
+
+@login_required(login_url="/")
+def view_report(request, selected_records=None):
+    if request.method == "POST":
+        response = HttpResponse(content_type="application/pdf")
+        response["Content-Disposition"] = 'attachment; filename="Report.pdf"'
+
+        doc = SimpleDocTemplate(response, pagesize=letter)
+        styles = getSampleStyleSheet()
+        story = []
+
+        title_style = styles["Title"]
+        title = "Health Records Report"
+        story.append(Paragraph(title, title_style))
+        story.append(Spacer(1, 18))
+
+        right_aligned_style = ParagraphStyle(
+            "RightAligned", parent=styles["Normal"], alignment=TA_RIGHT
+        )
+        current_date = datetime.now().strftime(DATE_FORMAT)
+
+        logo = "healthScore/static/HSlogo.jpg"
+        logo_img = Image(logo, width=128, height=40)
+        logo_and_date = [
+            [logo_img, Paragraph("Date: " + current_date, right_aligned_style)]
+        ]
+        logo_and_date = Table(logo_and_date)
+        logo_and_date.setStyle(
+            TableStyle(
+                [
+                    ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),
+                    ("ALIGN", (1, 0), (1, 0), "RIGHT"),
+                    ("BOTTOMPADDING", (0, 0), (-1, -1), 12),
+                ]
+            )
+        )
+        story.append(logo_and_date)
+        user_info = request.user
+        story.append(Paragraph("Name: " + user_info.name, styles["Normal"]))
+        try:
+            story.append(
+                Paragraph(
+                    "DOB: " + user_info.dob.strftime("%Y-%m-%d"), styles["Normal"]
+                )
+            )
+        except Exception:
+            story.append(Paragraph("DOB: " + "", styles["Normal"]))
+
+        story.append(Paragraph("BloodGroup: " + user_info.bloodGroup, styles["Normal"]))
+        story.append(Paragraph("Email: " + user_info.email, styles["Normal"]))
+        story.append(Paragraph("Contact: " + user_info.contactInfo, styles["Normal"]))
+        story.append(Paragraph("Address: " + user_info.address, styles["Normal"]))
+        story.append(Spacer(1, 12))
+
+        table_data = [
+            [
+                Paragraph("Reason for Visit"),
+                Paragraph("Visit Details"),
+                Paragraph("Healthcare Worker"),
+                Paragraph("Healthcare Facility"),
+                Paragraph("Address"),
+                Paragraph("Date"),
+                Paragraph("Properties"),
+            ],
+        ]
+
+        selected_record_ids = selected_records
+        if not selected_records:
+            selected_record_ids = request.POST.getlist("record_ids")
+        for record_id in selected_record_ids:
+            row = []
+            record = HealthRecord.objects.get(id=record_id)
+            appointment_pro = record.appointmentId.properties
+            appointment_properties = json.loads(appointment_pro)
+            appointment_name = record.appointmentId.name
+            appointment_name_para = Paragraph(appointment_name)
+            row.append(appointment_name_para)
+
+            appointment_type = appointment_properties.get("type", "Unknown")
+            appointment_type_para = Paragraph(appointment_type)
+            row.append(appointment_type_para)
+
+            doctor_name = HospitalStaff.objects.get(id=record.doctorID).name
+            doctor_name_para = Paragraph(doctor_name)
+            row.append(doctor_name_para)
+
+            hospital_name = Hospital.objects.get(id=record.hospitalID).name
+            hospital_name_para = Paragraph(hospital_name)
+            row.append(hospital_name_para)
+
+            hospital_addr = Hospital.objects.get(id=record.hospitalID).address
+            hospital_addr_para = Paragraph(hospital_addr)
+            row.append(hospital_addr_para)
+
+            updated = record.updatedAt.strftime(DATE_FORMAT)
+            updated_para = Paragraph(updated)
+            row.append(updated_para)
+
+            temp_row = []
+            for rec, val in appointment_properties.items():
+                temp_row.append(Paragraph(str(rec).capitalize() + " :   " + str(val)))
+            row.append(temp_row)
+
+            table_data.append(row)
+
+        page_width, page_height = letter
+        left_margin = right_margin = 50
+        effective_page_width = page_width - (left_margin + right_margin)
+
+        col_widths = [
+            effective_page_width * 0.1,  # Reason for Visit
+            effective_page_width * 0.15,  # Visit Details
+            effective_page_width * 0.15,  # Healthcare Worker
+            effective_page_width * 0.15,  # Healthcare Facility
+            effective_page_width * 0.15,  # Address
+            effective_page_width * 0.15,  # Date
+            effective_page_width * 0.25,  # Properties
+        ]
+        table = Table(table_data, colWidths=col_widths)
+
+        table_style = TableStyle(
+            [
+                ("BACKGROUND", (0, 0), (-1, 0), colors.grey),
+                ("TEXTCOLOR", (0, 0), (-1, 0), colors.whitesmoke),
+                ("ALIGN", (0, 0), (-1, -1), "CENTER"),
+                ("FONTNAME", (0, 0), (-1, 0), "Helvetica-Bold"),
+                ("BOTTOMPADDING", (0, 0), (-1, 0), 12),
+                ("BACKGROUND", (0, 1), (-1, -1), colors.white),
+                ("GRID", (0, 0), (-1, -1), 1, colors.black),
+                ("TEXTCOLOR", (0, 0), (-1, -1), colors.black),
+                ("VALIGN", (0, 0), (-1, -1), "TOP"),
+                ("WORDWRAP", (0, 0), (-1, -1), "CJK"),
+            ]
+        )
+
+        table.setStyle(table_style)
+        story.append(table)
+
+        doc.build(story)
+        return response
+
+
+def get_record(request, rec_id):
+    healthRecordList = list(HealthRecord.objects.filter(id=rec_id).values())
+
+    return JsonResponse({"data": json.dumps(healthRecordList[0], default=str)})
+
+
+@login_required(login_url="/")
+def view_health_history(request):
+    # Create a new QueryDict object with the desired parameters: fetch only approved records for health history page
+    if request.user.is_patient:
+        updated_params = request.GET.copy()
+        updated_params["record_status"] = "approved"
+
+        # Update request.GET with the modified QueryDict
+        request.GET = updated_params
+
+        zipped_details = get_health_history_details(request=request)
+        return render(request, "view_history.html", {"zipped_details": zipped_details})
+    return redirect("homepage")
+
+
